pyproject.toml
[project]
name = "rp-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "weaviate-client (>=4.15.0,<5.0.0)",
    "openai (>=1.82.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "pdfminer-six (==20250327)",
    "pi-heif (>=0.22.0,<0.23.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "unstructured-inference (>=1.0.2,<2.0.0)",
    "unstructured-pytesseract (>=0.3.15,<0.4.0)"
]

[tool.poetry]
packages = [{include = "rp_poetry", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
